moved mergeAggrs into another file
